ai_service/requirements.txt
fastapi
uvicorn
python-dotenv
gitpython
sentence-transformers
faiss-cpu
requests
numpy